# backend/MontagGo.API/Request/MyRegisterRequest.cs
using System.ComponentModel.DataAnnotations;

namespace MontagGo.API.Request
{
    public class MyRegisterRequest
    {
        [Required]
        public string Username { get; set; }

        [Required, EmailAddress]
        public string Email { get; set; }

        [Required, MinLength(6)]
        public string Password { get; set; }

        [Required, Compare("Password")]
        public string ConfirmPassword { get; set; }
    }
}